Seeder data adderd, fixed the thing that lets the route details refresh, still have to add commentary to locations

+++ Examen_Opdracht_NET_Advanced/Models/Route.cs
@@ -11,8 +11,8 @@ namespace Examen_Opdracht_NET_Advanced.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Length { get; set; }
-        public int Duration { get; set; }
+        public double Length { get; set; }
+        public double Duration { get; set; }
 
         public List<Progres>? ProgresesRoute { get; set; }
     }
